+++ class/utils/CdarHandler.class.php
@@ -379,8 +379,9 @@ class CdarHandler
 		} else {
 			// We issue the status as the BUYER of a supplier invoice, and it goes back to the vendor
 			$CdarIssuerTradeParty = [
-				'GlobalID' => $mysocGlobalID, // GlobalID of CDAR SENDER
-				'RoleCode' => CdarHandler::ROLE_BY
+				'GlobalID'  => $mysocGlobalID, // GlobalID of CDAR SENDER
+				'SchemeID'  => CdarHandler::SCHEME_SIREN_0002,
+				'RoleCode'  => CdarHandler::ROLE_BY
 			];
 
 			$CdarRecipientTradeParty = [